+++ includes/frontend/class-header.php
@@ -6,7 +6,7 @@
  * @copyright  WebMan Design, Oliver Juhas
  *
  * @since    2.0.0
- * @version  2.4.1
+ * @version  2.4.4
  *
  * Contents:
  *
@@ -46,6 +46,7 @@ class Modern_Header {
 
 						add_action( 'tha_html_before', __CLASS__ . '::doctype' );
 
+						add_action( 'wp_head', __CLASS__ . '::charset', 0 );
 						add_action( 'wp_head', __CLASS__ . '::head', 1 );
 						add_action( 'wp_head', __CLASS__ . '::head_pingback', 1 );
 						add_action( 'wp_head', __CLASS__ . '::head_chrome_color', 1 );
@@ -111,6 +112,23 @@ class Modern_Header {
 				echo '<!doctype html>';
 
 		} // /doctype
+
+
+
+		/**
+		 * Meta charset.
+		 *
+		 * @since  2.4.4
+		 *
+		 * @return  void
+		 */
+		public static function charset() {
+
+			// Output
+
+				echo '<meta charset="' . esc_attr( get_bloginfo( 'charset' ) ) . '">' . PHP_EOL;
+
+		} // /charset